v.1.0 finished. Additional if statements added for conflict prevention.

+++ src/main/java/eventRegister/CategoryController.java
@@ -1,12 +1,11 @@
 package eventRegister;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.ArrayList;
+import java.util.List;
 
 @Controller
 public class CategoryController {
@@ -25,8 +24,14 @@ public class CategoryController {
 
     @PostMapping("/categories")
     public String create(@RequestParam String category_name) {
+        boolean exists = doesCategoryExist(category_name);
+        if(exists == false){
+            this.categoryRepository.save(new Category(category_name, new ArrayList<>()));
+        } else {
+            //It'd be better to return an informative page at this point
+            System.out.println("Category exists already!");
+        }
 
-        this.categoryRepository.save(new Category(category_name, new ArrayList<>()));
         return "redirect:/categories";
     }
 
@@ -39,7 +44,15 @@ public class CategoryController {
 
     @PostMapping("/categories/delete")
     public String delete(@RequestParam long id){
-        this.categoryRepository.deleteById(id);
+        List<Category> categories = this.categoryRepository.findAll();
+        Category cat = this.categoryRepository.getOne(id);
+        int eventQty = cat.getEvents().size();
+        if(eventQty == 0){
+            this.categoryRepository.deleteById(id);
+        }
+        else{
+            System.out.println("Cannot delete category as " + eventQty + " events belong to the category.");
+        }
         return "redirect:/categories";
     }
 
@@ -49,12 +62,30 @@ public class CategoryController {
 
         //Pretty lazy and not the best practice.
         if(!category_name.isEmpty()){
-            categoryToUpdate.setCategory_name(category_name);
+            boolean exists = doesCategoryExist(category_name);
+            if(exists == false){
+                categoryToUpdate.setCategory_name(category_name);
+            } else {
+                //It'd be better to return an informative page at this point
+                System.out.println("Category exists already!");
+            }
         } else{
+            //Update commit takes place but name changes the same, a workaround atm.
             categoryToUpdate.setCategory_name(categoryToUpdate.getCategory_name());
         }
-
         this.categoryRepository.save(categoryToUpdate);
         return "redirect:/categories";
+    }
+
+    //Needed by create and update methods to dismiss conflicts with unique category name restriction
+    public boolean doesCategoryExist(String proposedCategoryName){
+        boolean exists = false;
+        List<Category> categories = this.categoryRepository.findAll();
+        for(int i=0; i<categories.size(); i++){
+            if(categories.get(i).getCategory_name().toLowerCase().equals(proposedCategoryName.toLowerCase())){
+                exists = true;
+            }
+        }
+        return exists;
     }
 }
